refactor: 공통유틸 수정
